rename plot.Plot to project.Project
