+++ docs/Tutorial/datamodel.tex
@@ -1,0 +1,12 @@
+pygrace was designed so that, ``anything you can do in xmgrace, you can
+do in pygrace.''  Figure \ref{crowdiagram} is a diagram showing the
+basic organization of the pygrace structure --- it basically mirrors
+the xmgrace file format structure.  One thing to note is that Drawing
+Objects can be associated with either a Project or a Graph.  This is
+important in cases where one want to, e.g., create a custom data
+symbol and have the symbol show up in the coordinate system of a
+graph.
+
+\noindent
+For a more detailed diagram, look at the file {\tt
+  pygrace/documentation/diagram.pdf}.
