+++ docs/Tutorial/introduction.tex
@@ -1,7 +1,7 @@
-PyGrace is a Python package used for creating Grace files. Grace is a
-simple 2-D plotting program.  PyGrace is intended to:
+pygrace is a Python package used for creating xmgrace project files. xmgrace is
+a simple 2-D plotting program.  pygrace is intended to:
 \begin{itemize}
-\item enable Grace ({\tt .agr} format) file creation within Python.
+\item enable xmgrace ({\tt .agr} format) file creation within Python.
 \item assist in the automated creation of many figures for use in
   large-scale, exploratory data analysis.
 \item encourage code reuse, by offering an object-oriented structure.
@@ -9,17 +9,17 @@ simple 2-D plotting program.  PyGrace is intended to:
   journal articles.
 \end{itemize}
 Visualizing data is often an integral part of the data analysis
-process, however, \textbf{PyGrace is explicity intended not to perform
+process, however, \textbf{pygrace is explicity intended not to perform
   any type of data analysis}.  Not even computing an average.
 Especially not performing a linear regression.  This principle
-underlies much of the structure of PyGrace, with the purpose of not
+underlies much of the structure of pygrace, with the purpose of not
 allowing users to adopt a ``black box'' approach to analysis.  While
 visualizing data should be as painless as possible, the pain required
 to understand the analysis process is considered necessary, or even
 desirable (if you're into that sort of thing).
 
-This tutorial is meant to help with installation of PyGrace, and give
-new users an idea of how PyGrace is structured.  After reading this
-brief document, the next step for learning to use PyGrace is to go
-through the examples that are located in the {\tt PyGrace/Examples}
+This tutorial is meant to help with installation of pygrace, and give
+new users an idea of how pygrace is structured.  After reading this
+brief document, the next step for learning to use pygrace is to go
+through the examples that are located in the {\tt pygrace/examples}
 directory.
