+++ docs/Tutorial/main.tex
@@ -7,7 +7,7 @@
 
 \newenvironment{command}[0]{\begin{flushleft} \tt}{\end{flushleft}}
 
-\title{PyGrace Tutorial}
+\title{pygrace Tutorial}
 \date{}
 \begin{document}
 \maketitle
