--- documentation/Tutorial/datamodel.tex
@@ -1,12 +1,0 @@
-PyGrace was designed so that, ``anything you can do in Grace, you can
-do in PyGrace.''  Figure \ref{crowdiagram} is a diagram showing the
-basic organization of the PyGrace structure --- it basically mirrors
-the Grace file format structure.  One thing to note is that Drawing
-Objects can be associated with either a Grace or a Graph.  This is
-important in cases where one want to, e.g., create a custom data
-symbol and have the symbol show up in the coordinate system of a
-graph.
-
-\noindent
-For a more detailed diagram, look at the file {\tt
-  PyGrace/Documentation/diagram.pdf}.
