--- documentation/arXiv/acknowledgements.tex
@@ -1,2 +1,0 @@
-This product includes color specifications and designs developed by
-Cynthia Brewer (\url{http://colorbrewer.org/}).
