--- documentation/arXiv/conclusion.tex
@@ -1,2 +1,0 @@
-Conclusion.
-
