--- documentation/arXiv/examples.tex
@@ -1,1 +1,0 @@
-Examples.
