--- documentation/arXiv/introduction.tex
@@ -1,1 +1,0 @@
-Intro.
