+++ examples/00_helloworld.py
@@ -1,9 +1,9 @@
-from pygrace.plot import Plot
+from pygrace.project import Project
 
-# make an instance of the Grace class
-grace = Plot()
+# make an instance of the Project class
+grace = Project()
 
-# add a Graph as a "child" of the grace instance
+# add a Graph as a "child" of the Project instance
 graph = grace.add_graph()
 graph.title.text = 'Hello, world!'
 
@@ -12,5 +12,5 @@ graph.title.text = 'Hello, world!'
 data = [(0, 0), (0.5, 0.75), (1, 1)]
 dataset = graph.add_dataset(data)
 
-# print the grace to a file (.agr format)
-grace.write_file('00_helloworld.agr')
+# print the Project to a file (.agr format)
+grace.saveall('00_helloworld.agr')
