+++ examples/05_colorplot.py
@@ -1,23 +1,23 @@
 # This example demonstrates how to use a color plot.
 #
 # One important feature of color plots is that drawing objects in
-# XMGrace are plotted above the frame of a graph.  As a result, we use
-# the Grace.clone_graph method to copy the frame of the graph and make
+# xmgrace are plotted above the frame of a graph.  As a result, we use
+# the grace.clone_graph method to copy the frame of the graph and make
 # things look pretty.
 #
 # Another important feature of color plots is that specifying a
 # particular world coordinate can automatically place drawing objects
 # outside of the domain of the graph.
-# Graph.remove_extraworld_drawing_objects fixes this problem by
+# graph.remove_extraworld_drawing_objects fixes this problem by
 # removing all of the drawing objects that appear in the world
 # coordinates and are out of bounds.
 #
-# It also demonstrates how to use a PyGrace.Style to format figures in
+# It also demonstrates how to use a pygrace.style to format figures in
 # a customized format
 
 import random
 
-from pygrace.plot import Plot
+from pygrace.project import Project
 from pygrace.colors import ColorBrewerScheme
 from pygrace.axis import LINEAR_SCALE, LOGARITHMIC_SCALE
 from pygrace.drawing_objects import DrawBox
@@ -41,13 +41,13 @@ for pdf in pdfs:
     if pdf<minpdf and pdf>0.0:
         minpdf=pdf
 
-# instantiate a sweet figgy fig
-colors = ColorBrewerScheme("YlOrBr",n=253) # this is the maximum number of colors
+# instantiate a figure
+colors = ColorBrewerScheme("YlOrBr",n=253) # this is the max number of colors
 
-# you can change the opacity percent of a colorscheme if you want:
+# change the opacity percent of a colorscheme
 # colors.change_opacity(20, exclude_black=False)
 
-grace = Plot(colors=colors)
+grace = Project(colors=colors)
 
 # add a colorbar
 colorbar = grace.add_graph(ElLogColorBar,domain=(minpdf,maxpdf),
@@ -69,7 +69,7 @@ colorbar.set_view(0.95,0.2,1.0,0.9)
 
 # label axes
 colorbar.set_label("Probability density, p(x,y)")
-graph.xaxis.label.text = "Mike is better than Dean"
+graph.xaxis.label.text = "Model"
 graph.yaxis.label.text = "Truth"
 
 # all of the autoscale features work, too
@@ -84,5 +84,5 @@ graph.autoticky()
 # get rid of points that are out of bounds
 graph.remove_extraworld_drawing_objects()
 
-# print out the grace
-grace.write_file("05_colorplot.agr")
+# print out the Project
+grace.saveall("05_colorplot.agr")
