+++ examples/12_cheatsheet.py
@@ -1,7 +1,7 @@
-from pygrace.plot import Plot
+from pygrace.project import Project
 from pygrace.drawing_objects import DrawBox, DrawText, DrawLine, DrawEllipse
 
-grace = Plot()
+grace = Project()
 graph = grace.add_graph()
 graph.add_dataset([(0, 0), (1, 1)])
 
@@ -10,4 +10,4 @@ graph.add_drawing_object(DrawText)
 graph.add_drawing_object(DrawLine)
 graph.add_drawing_object(DrawEllipse)
 
-grace.write_cheatsheet('cheatsheet.tex')
+grace.cheatsheet('cheatsheet.tex')
