+++ pygrace/extensions/multi_grace.py
@@ -1,2 +1,2 @@
 # preserving backward compatibility with PyGrace
-from .multi_plot import MultiPlot as MultiGrace
+from .multi_plot import MultiPlotProject as MultiGrace
