+++ pygrace/grace.py
@@ -1,2 +1,2 @@
 # preserving backward compatibility with PyGrace
-from .plot import Plot as Grace
+from .project import Project as Grace
